refactor: rename occupation components
